Fix Narrowing several regions within the same file
If you narrow several regions within the same file and want to write
several changes independently, you should set the g:nrrw_rgn_protect
variable to 'n'.
However if you use it, NarrowRegion would not take into account, that
line numbers could have been changed, so this needs to be adjusted as
well as that the highlighting needed to be adjusted as well.

This should work now.